Make all paths to top-level assets relative.
This is annoying, but currently MyST fails to correctly resolve absolute paths as starting at the root of the project. So for now, for images to be found, they need to have a full relative path from the source of their use to the actual image on disk.

See https://github.com/jupyter-book/mystmd/issues/1687